udpate supbased data apply new pages
