+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@mdi/font": "^7.1.96",
+    "@supabase/supabase-js": "^2.12.0",
     "axios": "^1.3.4",
     "pinia": "^2.0.28",
     "vite-plugin-vuetify": "^1.0.1",
