+++ src/components/products/products-page.vue
@@ -1,0 +1,12 @@
+<template>
+    <section class="w-100 h-100"></section>
+</template>
+
+<script lang="ts">
+
+    export default ({
+        data() {
+
+        }
+    });
+</script>
