+++ src/styles/company.css
@@ -1,0 +1,13 @@
+.company-section {
+    display: flex;
+}
+
+.member-details {
+    width: 0;
+    height: 100%;
+}
+
+.member-list {
+    width: 100%;
+    height: 100%;
+}
